+++ src/game.rs
@@ -94,6 +94,17 @@ pub fn creeps() -> JsHashMap<String, Creep> {
     Game::creeps().into()
 }
 
+/// Get a [`JsHashMap<JsString, Creep>`] with all of your creeps, which has
+/// creep names as keys.
+///
+/// Note that newly spawned creeps are immediately added when spawned, but will
+/// not have an id until the following tick.
+///
+/// [Screeps documentation](https://docs.screeps.com/api/#Game.creeps)
+pub fn creeps_jsstring() -> JsHashMap<JsString, Creep> {
+    Game::creeps().into()
+}
+
 /// Get a [`JsHashMap<String, Flag>`] with all of your flags, which has flag
 /// names as keys.
 ///
@@ -102,11 +113,27 @@ pub fn flags() -> JsHashMap<String, Flag> {
     Game::flags().into()
 }
 
+/// Get a [`JsHashMap<JsString, Flag>`] with all of your flags, which has flag
+/// names as keys.
+///
+/// [Screeps documentation](https://docs.screeps.com/api/#Game.flags)
+pub fn flags_jsstring() -> JsHashMap<JsString, Flag> {
+    Game::flags().into()
+}
+
 /// Get a [`JsHashMap<String, AccountPowerCreep>`] with all of your power
 /// creeps, which has power creep names as keys.
 ///
 /// [Screeps documentation](https://docs.screeps.com/api/#Game.powerCreeps)
 pub fn power_creeps() -> JsHashMap<String, AccountPowerCreep> {
+    Game::power_creeps().into()
+}
+
+/// Get a [`JsHashMap<JsString, AccountPowerCreep>`] with all of your power
+/// creeps, which has power creep names as keys.
+///
+/// [Screeps documentation](https://docs.screeps.com/api/#Game.powerCreeps)
+pub fn power_creeps_jsstring() -> JsHashMap<JsString, AccountPowerCreep> {
     Game::power_creeps().into()
 }
 
@@ -131,6 +158,14 @@ pub fn rooms() -> JsHashMap<RoomName, Room> {
 ///
 /// [Screeps documentation](https://docs.screeps.com/api/#Game.spawns)
 pub fn spawns() -> JsHashMap<String, StructureSpawn> {
+    Game::spawns().into()
+}
+
+/// Get a [`JsHashMap<JsString, StructureSpawn>`] with all of your spawns, which
+/// has spawn names as keys.
+///
+/// [Screeps documentation](https://docs.screeps.com/api/#Game.spawns)
+pub fn spawns_jsstring() -> JsHashMap<JsString, StructureSpawn> {
     Game::spawns().into()
 }