+++ src/objects/impls/owned_structure.rs
@@ -1,4 +1,3 @@
-use js_sys::JsString;
 use wasm_bindgen::prelude::*;
 
 use crate::{
@@ -63,13 +62,7 @@ extern "C" {
     #[wasm_bindgen]
     pub type Owner;
 
-    #[wasm_bindgen(method, getter = username)]
-    fn username_internal(this: &Owner) -> JsString;
-}
-
-impl Owner {
     /// The name of the player that owns this object.
-    pub fn username(&self) -> String {
-        Self::username_internal(self).into()
-    }
+    #[wasm_bindgen(method, getter)]
+    pub fn username(this: &Owner) -> String;
 }