+++ src/objects/impls/structure_spawn.rs
@@ -28,11 +28,17 @@ extern "C" {
     #[wasm_bindgen(method, setter)]
     pub fn set_memory(this: &StructureSpawn, val: &JsValue);
 
-    /// The spawn's name as an owned reference to a [`JsString`].
+    /// The spawn's name as a [`String`].
     ///
     /// [Screeps documentation](https://docs.screeps.com/api/#StructureSpawn.name)
     #[wasm_bindgen(method, getter)]
-    pub fn name(this: &StructureSpawn) -> JsString;
+    pub fn name(this: &StructureSpawn) -> String;
+
+    /// The spawn's name as a [`JsString`].
+    ///
+    /// [Screeps documentation](https://docs.screeps.com/api/#StructureSpawn.name)
+    #[wasm_bindgen(method, getter = name)]
+    pub fn name_jsstring(this: &StructureSpawn) -> JsString;
 
     /// Information about the spawning creep, if one is currently being spawned.
     ///