+++ src/raw_memory.rs
@@ -48,10 +48,19 @@ extern "C" {
 
 /// Get a [`JsHashMap<u8, String>`] with all of the segments requested on
 /// the previous tick, with segment numbers as keys and segment data in
-/// [`JsString`] form as values.
+/// [`String`] form as values.
 ///
 /// [Screeps documentation](https://docs.screeps.com/api/#RawMemory.segments)
 pub fn segments() -> JsHashMap<u8, String> {
+    RawMemory::segments().into()
+}
+
+/// Get a [`JsHashMap<u8, JsString>`] with all of the segments requested on
+/// the previous tick, with segment numbers as keys and segment data in
+/// [`JsString`] form as values.
+///
+/// [Screeps documentation](https://docs.screeps.com/api/#RawMemory.segments)
+pub fn segments_jsstring() -> JsHashMap<u8, JsString> {
     RawMemory::segments().into()
 }